Converted backing fields to double, and implemented Magnitude

+++ src/raytracer/Point.cs
@@ -2,6 +2,6 @@ namespace codeclimber.raytracer
 {
     public class Point: Tuple
     {
-        public Point(float x, float y, float z):base(x,y,z,1.0f) { }
+        public Point(double x, double y, double z):base(x,y,z,1.0f) { }
     }
 }
